Updated overlays - Added new overlays - Moved functions over to it's own module: "overlays.py" - Added temporary c|overlays command to list existing overlays
