+++ .gitignore
@@ -109,5 +109,6 @@ pip-selfcheck.json
 
 config.json
 discord.log
+dbcompile.py
 *.db
 *.desktop
